Change logging level to DEBUG for active current and cast to int16_t in getActiveCurrent to fix overflow error

+++ zoe2_hardware/src/commander.cpp
@@ -311,7 +311,7 @@ int Command::getActiveCurrent(int* current, unsigned int can_id) {
     return -2;
   }
 
-  *current = intFromDataBigEndian(frame.data, 0, 2);
+  *current = (int16_t) intFromDataBigEndian(frame.data, 0, 2);
 
   RCLCPP_DEBUG(rclcpp::get_logger("COB"), "Active current: %d", *current);
